pyproject.toml
[project]
name = "regulatory-newsfeed-poc"
version = "0.1.0"
description = "Regulatory Newsfeed Drafting Automation (POC) — Water sector"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    # Orchestration
    "prefect>=2.19.0,<3.0.0",

    # HTTP / scraping
    "httpx[http2]>=0.27.0",
    "playwright>=1.45.0",
    "selectolax>=0.3.21",
    "trafilatura>=1.12.0",
    "feedparser>=6.0.11",
    "lxml>=5.2.0",

    # Markdown / parsing
    "markdown-it-py>=3.0.0",

    # PDF
    "pdfplumber>=0.11.0",
    "ocrmypdf>=16.5.0",

    # Watchers / notifications
    "watchdog>=4.0.0",
    "imap-tools>=1.7.0",
    "plyer>=2.1.0",

    # Persistence
    "SQLAlchemy>=2.0.30",

    # Resilience
    "tenacity>=9.0.0",

    # Config / validation / secrets
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",

    # Templating
    "Jinja2>=3.1.4",

    # Dates
    "python-dateutil>=2.9.0",
    "dateparser>=1.2.0",

    # LLM providers
    "boto3>=1.35.0",
    "openai>=1.40.0",

    # Word output
    "python-docx>=1.1.2",

    # Review UI
    "streamlit>=1.38.0",

    # Logging
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "bandit>=1.7.9",
    "types-PyYAML>=6.0.12",
    "types-python-dateutil>=2.9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "PL"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "lxml",
    "lxml.*",
    "boto3",
    "boto3.*",
    "botocore",
    "botocore.*",
    "feedparser",
    "trafilatura",
    "selectolax.*",
    "pdfplumber",
    "ocrmypdf",
    "dateparser",
    "plyer",
    "plyer.*",
    "imap_tools",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "unit: unit tests (fast, no IO)",
    "integration: integration tests (may touch disk / network)",
    "live: hits real external APIs — skipped by default",
]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src/newsfeed"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80
